seperated relationship popover copm from popoverComponents and added some styles
