+++ silzila-frontend/src/redux/Dataset/datasetActions.js
@@ -26,9 +26,24 @@ export const removeArrows = (pl) => {
 	return { type: "REMOVE_ARROWS", payload: pl };
 };
 export const resetState = () => {
-    return { type: "RESET_STATE" };
+	return { type: "RESET_STATE" };
 };
 
 // =============================================================================
 // Actions from Canvas
 // =============================================================================
+export const addArrows = (arrow) => {
+	return { type: "ADD_ARROWS", payload: arrow };
+};
+export const clickOnArrow = (payload) => {
+	return { type: "CLICK_ON_ARROW", payload: payload };
+};
+export const setArrowType = (payload) => {
+	return { type: "SET_ARROW_TYPE", payload: payload };
+};
+export const setArrows = (pl) => {
+	return { type: "SET_ARROWS", payload: pl };
+};
+export const resetArrows = () => {
+	return { type: "RESET_ARROWS_ARRAY" };
+};
